use size_t for sizes

+++ cnpy.cpp
@@ -56,7 +56,7 @@ template<> std::vector<char>& cnpy::operator+=(std::vector<char>& lhs, const cha
     return lhs;
 }
 
-void cnpy::parse_npy_header(FILE* fp, unsigned int& word_size, unsigned int*& shape, unsigned int& ndims, bool& fortran_order) {  
+void cnpy::parse_npy_header(FILE* fp, size_t& word_size, size_t*& shape, size_t& ndims, bool& fortran_order) {  
     char buffer[256];
     size_t res = fread(buffer,sizeof(char),11,fp);       
     if(res != 11)
@@ -76,8 +76,8 @@ void cnpy::parse_npy_header(FILE* fp, unsigned int& word_size, unsigned int*& sh
     std::string str_shape = header.substr(loc1+1,loc2-loc1-1);
     if(str_shape[str_shape.size()-1] == ',') ndims = 1;
     else ndims = std::count(str_shape.begin(),str_shape.end(),',')+1;
-    shape = new unsigned int[ndims];
-    for(unsigned int i = 0;i < ndims;i++) {
+    shape = new size_t[ndims];
+    for(size_t i = 0;i < ndims;i++) {
         loc1 = str_shape.find(",");
         shape[i] = atoi(str_shape.substr(0,loc1).c_str());
         str_shape = str_shape.substr(loc1+1);
@@ -98,7 +98,7 @@ void cnpy::parse_npy_header(FILE* fp, unsigned int& word_size, unsigned int*& sh
     word_size = atoi(str_ws.substr(0,loc2).c_str());
 }
 
-void cnpy::parse_zip_footer(FILE* fp, unsigned short& nrecs, unsigned int& global_header_size, unsigned int& global_header_offset)
+void cnpy::parse_zip_footer(FILE* fp, unsigned short& nrecs, size_t& global_header_size, size_t& global_header_offset)
 {
     std::vector<char> footer(22);
     fseek(fp,-22,SEEK_END);
@@ -122,16 +122,16 @@ void cnpy::parse_zip_footer(FILE* fp, unsigned short& nrecs, unsigned int& globa
 }
 
 cnpy::NpyArray load_the_npy_file(FILE* fp) {
-    unsigned int* shape;
-    unsigned int ndims, word_size;
+    size_t* shape;
+    size_t ndims, word_size;
     bool fortran_order;
     cnpy::parse_npy_header(fp,word_size,shape,ndims,fortran_order);
     unsigned long long size = 1; //long long so no overflow when multiplying by word_size
-    for(unsigned int i = 0;i < ndims;i++) size *= shape[i];
+    for(size_t i = 0;i < ndims;i++) size *= shape[i];
 
     cnpy::NpyArray arr;
     arr.word_size = word_size;
-    arr.shape = std::vector<unsigned int>(shape,shape+ndims);
+    arr.shape = std::vector<size_t>(shape,shape+ndims);
     arr.data = new char[size*word_size];    
     arr.fortran_order = fortran_order;
     size_t nread = fread(arr.data,word_size,size,fp);
@@ -219,7 +219,7 @@ cnpy::NpyArray cnpy::npz_load(std::string fname, std::string varname) {
         }
         else {
             //skip past the data
-            unsigned int size = *(unsigned int*) &local_header[22];
+            size_t size = *(unsigned int*) &local_header[22];
             fseek(fp,size,SEEK_CUR);
         }
     }
