+++ example1.cpp
@@ -16,7 +16,7 @@ int main()
     for(int i = 0;i < Nx*Ny*Nz;i++) data[i] = std::complex<double>(rand(),rand());
 
     //save it to file
-    const unsigned int shape[] = {Nz,Ny,Nx};
+    const size_t shape[] = {Nz,Ny,Nx};
     cnpy::npy_save("arr1.npy",data,shape,3,"w");
 
     //load it into a new array
@@ -36,7 +36,7 @@ int main()
     //non-array variables are treated as 1D arrays with 1 element
     double myVar1 = 1.2;
     char myVar2 = 'a';
-    unsigned int shape2[] = {1};
+    size_t shape2[] = {1};
     cnpy::npz_save("out.npz","myVar1",&myVar1,shape2,1,"w"); //"w" overwrites any existing file
     cnpy::npz_save("out.npz","myVar2",&myVar2,shape2,1,"a"); //"a" appends to the file we created above
     cnpy::npz_save("out.npz","arr1",data,shape,3,"a"); //"a" appends to the file we created above
